🏷️(app-desk) move Role type to Team
We need the Role type in Team as well.
Better to move it to the highest level.
