+++ src/frontend/apps/desk/src/features/teams/api/index.ts
@@ -1,4 +1,3 @@
-export * from './types';
 export * from './useCreateTeam';
 export * from './useTeam';
 export * from './useTeams';
