+++ src/frontend/apps/desk/src/features/teams/index.tsx
@@ -1,2 +1,4 @@
-export * from './components';
 export * from './api';
+export * from './components';
+export * from './types';
+export * from './store';
