+++ src/frontend/apps/desk/src/features/teams/store/index.ts
@@ -1,0 +1,1 @@
+export * from './useTeamsStore';
